car_dealership/requirements.txt
Django==4.0.2
django-money
django_countries
psycopg2-binary==2.9.1
tzdata
djangorestframework_simplejwt
djangorestframework
drf_yasg
django-debug-toolbar==3.2.4
django-filter